+++ src/entities/home/ui/homeArticlePreview/index.css
