+++ src/index.html
@@ -8,8 +8,7 @@
     <link rel="stylesheet" href="./app/styles/variables.css" />
   </head>
   <body>
-    <div id="app">
-      <script type="module" src="./app/routes/route.ts"></script>
-    </div>
+    <div id="app"></div>
+    <script type="module" src="./app/routes/route.ts"></script>
   </body>
 </html>
